Optimize hero image - Increased performance by 26% using Nextjs Image component for the hero bg - Updated logo image from svg to png for Next Image optimization - Imported new size AD images and renamed - Imported a new size Header Img
