+++ data/AdData.ts
@@ -1,16 +1,16 @@
 const AdData = [
   {
-    img: '/static/SBImg.png',
+    img: '/static/ADImg1.png',
     heading: 'Celebrate Sports',
     desc: 'Capturing Athletic Spirit in Every Snap! Step into our sports-themed photo booth & celebrate your teams victories 📸🏆',
   },
   {
-    img: '/static/GalentinesImg.png',
+    img: '/static/ADImg2.png',
     heading: 'Celebrate Love',
     desc: 'Where Love Sparks Forever in Every Frame! Step into our romantic-themed photo booth & let your love story unfold 💕📸',
   },
   {
-    img: '/static/AdEvents.png',
+    img: '/static/ADImg3.png',
     heading: 'Corporate Events',
     desc: 'Elevating Corporate Events with Memorable Moments! Hire us for your next event!',
   },
